moved type defs into types.ts
