+++ src/client/datascience/serviceRegistry.ts
@@ -35,13 +35,14 @@ import { CodeWatcher } from './editor-integration/codewatcher';
 import { Decorator } from './editor-integration/decorator';
 import { HoverProvider } from './editor-integration/hoverProvider';
 import { DataScienceErrorHandler } from './errorHandler/errorHandler';
-import { ExportFormat, ExportManager, IExport, IExportManager } from './export/exportManager';
+import { ExportManager } from './export/exportManager';
 import { ExportManagerDependencyChecker } from './export/exportManagerDependencyChecker';
 import { ExportManagerFileOpener } from './export/exportManagerFileOpener';
 import { ExportManagerFilePicker, IExportManagerFilePicker } from './export/exportManagerFilePicker';
 import { ExportToHTML } from './export/exportToHTML';
 import { ExportToPDF } from './export/exportToPDF';
 import { ExportToPython } from './export/exportToPython';
+import { ExportFormat, IExport, IExportManager } from './export/types';
 import { GatherListener } from './gather/gatherListener';
 import { GatherLogger } from './gather/gatherLogger';
 import { DebugListener } from './interactive-common/debugListener';
