+++ src/test/.vscode/settings.json
@@ -5,6 +5,8 @@
     "python.testing.nosetestArgs": [],
     "python.testing.pytestArgs": [],
     "python.testing.unittestArgs": [
+        "-s=./tests",
+        "-p=test_*.py",
         "-v",
         "-s",
         ".",
