+++ src/test/datascience/export/exportToPython.test.ts
@@ -7,7 +7,7 @@ import * as path from 'path';
 import { Uri } from 'vscode';
 import { IDocumentManager } from '../../../client/common/application/types';
 import { IFileSystem } from '../../../client/common/platform/types';
-import { ExportFormat, IExport } from '../../../client/datascience/export/exportManager';
+import { ExportFormat, IExport } from '../../../client/datascience/export/types';
 import { IExtensionTestApi } from '../../common';
 import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../constants';
 import { closeActiveWindows, initialize } from '../../initialize';
